feat: Updated to 1.21
Remade the project with the latest architectury project from the MC
development IntelliJ plugin. This just madde it easier to add
neoforge support, because I really can't be arsed to create the
project for it manually, and also it's pretty much a 1-file mod.

Only annoying thing about this port was figuring out why the
advanccements weren't loading - turns out that 1.21 changed the name
of the advancements data folder from "advancements" to "advancement".
Renaming it fixed it.

I've also dropped the embedded Fabric API, as pretty much every
fabric user is going to have it installed anyway, so it's just
unnessesary code duplication. Sure, it's less self-contained, but a
Fabric API dependency is pretty much a given.